requirements.txt
aiogram==3.0.0
APScheduler==3.9.1.post1
python-dotenv==1.0.0